+++ packages/nest/lib/index.ts
@@ -1,4 +1,4 @@
 export * from './decorators';
 export * from './utils';
-export * from './mediator.module.options';
+export * from './options';
 export * from './mediator.module';
