+++ packages/nest/lib/options/index.ts
@@ -1,0 +1,1 @@
+export * from './mediator.module.options';
